+++ basics/04_modules/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "modules"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
